(HTCONDOR-3731)  Derp.  We only test container jobs on platforms which support them.
