+++ src/condor_tests/test_cif_catalogs.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env pytest
 
 import pytest
-import subprocess
 
 import logging
 
@@ -22,6 +21,12 @@ from pathlib import Path
 import shutil
 
 import htcondor2
+
+from libcontainer import (
+    SingularityIsWorthy,
+    UserNamespacesFunctional,
+    SingularityIsWorking,
+)
 
 
 def try_shutil_chown( * p, ** v ):
@@ -583,52 +588,6 @@ def shadow_log_is_as_expected(the_condor, count, cf_xfers, cf_waits):
             the_condor, "Waiting for common files to be transferred"
         )
         assert common_transfer_waits == cf_waits
-
-
-# ---- Singularity checks -----------------------------------------------------
-# All stolen from `test_singularity_sif.py`, and which should probably be made
-# Ornithology fixtures.
-
-def SingularityIsWorthy():
-    result = subprocess.run("singularity --version", shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-    output = result.stdout.decode('utf-8')
-
-    logger.debug(output)
-    if "apptainer" in output:
-        return True
-
-    if "3." in output:
-        return True
-
-    return False
-
-
-def SingularityIsWorking():
-    result = subprocess.run("singularity exec -B/bin:/bin -B/lib:/lib -B/lib64:/lib64 -B/usr:/usr busybox.sif /bin/ls /", shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-    output = result.stdout.decode('utf-8')
-
-    logger.debug(output)
-
-    if result.returncode == 0:
-        return True
-    else:
-        return False
-
-
-# For the test to work, we need user namespaces to be working
-# and enough of them.  This is a race, but better to try
-# to test first.
-def UserNamespacesFunctional():
-    try:
-        result = subprocess.run(["unshare", "-U", "/bin/sh", "-c", "exit 7"])
-    except FileNotFoundError:
-        return False
-    if result.returncode == 7:
-        print("unshare seems to work correctly, proceeding with test\n")
-        return True
-    else:
-        print("unshare command failed, test cannot work, skipping test\n")
-        return False
 
 
 # ---- Tests ------------------------------------------------------------------
